+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for defining a project's verbose title (DOCSP-7988)
+
 ## [v0.2.0] - 2020-01-23
 
 ### Added